feat: script to fix localized files
When babel::render_book is run, it converts all the .qmd file to html and adds a prefix for the languages short code if several languages are specified.

The issue here is that in the doc generated for the languages it includes the index.html but when you try to select the new language on the rendered book, it references index.<language short code>.html instead of index.html

Taking Arabic as an example, the language shortcode would be 'ar'. The generated doc will be /ar/index.html, when i try to switch to Arabic version of the rendered website it'll goto http://<your host>/ar/index.ar instead of http://<your host>/ar/index or http://<your host>/ar

SOLUTION
This fix adds a index.<language short code>.html to the generated docs for each language